Add GitHub Actions CI, 0.0.5 Version Release.
fix:
* (core): Improved version.vim.
          Use popupwin instead of popup to check popup window support.

* (layers): Bracket display in NERDTree, conflict with plugin rainbow.
            Ctags/Gtags settings.

* (installing): Error when dir ~/.config not exist.

test:
* (actions): Add GitHub Actions CI.

+++ bootstrap.sh
@@ -193,10 +193,14 @@ create_symlinks() {
     lnif "$source_path/.vimrc"         "$target_path/.vimrc"
 
     if program_exists "nvim"; then
-        today=`date +%Y%m%d_%s`
-        for i in "$2/.config/nvim" "$2/.config/nvim/init.vim"; do
-                [ -e "$i" ] && [ ! -L "$i" ] && mv -v "$i" "$i.$today";
-        done
+        if [ ! -e "$2/.config" ]; then
+            mkdir -p "$2/.config"
+        else
+            today=`date +%Y%m%d_%s`
+            for i in "$2/.config/nvim" "$2/.config/nvim/init.vim"; do
+                    [ -e "$i" ] && [ ! -L "$i" ] && mv -v "$i" "$i.$today";
+            done
+        fi
         lnif "$2/.vim"                 "$target_path/.config/nvim"
         lnif "$source_path/.vimrc"     "$target_path/.config/nvim/init.vim"
         [ -L "$2/.vim/.vim" ] && rm "$2/.vim/.vim"
